@@ -1,162 +1,166 @@
 {
-    "name": "arenatime-client",
-    "version": "0.1.0",
-    "homepage": "https://stdev17.github.io/arenatime",
-    "private": true,
-    "dependencies": {
-        "@babel/core": "^7.5.5",
-        "@svgr/webpack": "4.1.0",
-        "athena-client": "^2.5.1",
-        "athena-express": "^5.1.0",
-        "aws-sdk": "^2.500.0",
-        "axios": "^0.19.0",
-        "babel-eslint": "10.0.1",
-        "babel-jest": "^21.3.0-beta.15",
-        "babel-loader": "8.0.5",
-        "babel-plugin-named-asset-import": "^0.3.2",
-        "babel-preset-react-app": "^9.0.0",
-        "bootstrap": "^4.3.1",
-        "camelcase": "^5.2.0",
-        "case-sensitive-paths-webpack-plugin": "2.2.0",
-        "chai": "^4.2.0",
-        "chai-as-promised": "^7.1.1",
-        "compression": "^1.7.4",
-        "css-loader": "2.1.1",
-        "dotenv": "6.2.0",
-        "dotenv-expand": "4.2.0",
-        "eslint-config-react-app": "^4.0.1",
-        "eslint-loader": "2.1.2",
-        "eslint-plugin-flowtype": "2.50.1",
-        "eslint-utils": "^1.4.2",
-        "fetch": "^1.1.0",
-        "file-loader": "3.0.1",
-        "file-type": "^12.1.0",
-        "form-data": "^2.5.0",
-        "fs": "0.0.1-security",
-        "fs-extra": "7.0.1",
-        "gh-pages": "^2.0.1",
-        "html-webpack-plugin": "4.0.0-beta.5",
-        "identity-obj-proxy": "3.0.0",
-        "is-mobile": "^2.1.0",
-        "is-wsl": "^1.1.0",
-        "istanbul-lib-instrument": "^3.3.0",
-        "jest": "^21.3.0-beta.15",
-        "jest-environment-jsdom-fourteen": "0.1.0",
-        "jest-resolve": "24.7.1",
-        "jest-watch-typeahead": "0.3.0",
-        "js-combinatorics": "^0.5.4",
-        "konva": "^3.4.1",
-        "lodash": "^4.17.14",
-        "lodash.template": "^4.5.0",
-        "mini-css-extract-plugin": "0.5.0",
-        "mixin-deep": "^2.0.1",
-        "moment": "^2.24.0",
-        "multimatch": "^4.0.0",
-        "node-gzip": "^1.1.2",
-        "nth-prime": "^1.0.4",
-        "optimize-css-assets-webpack-plugin": "5.0.1",
-        "pnp-webpack-plugin": "1.2.1",
-        "postcss-flexbugs-fixes": "4.1.0",
-        "postcss-loader": "3.0.0",
-        "postcss-normalize": "7.0.1",
-        "postcss-preset-env": "6.6.0",
-        "postcss-safe-parser": "4.0.1",
-        "react": "^16.8.6",
-        "react-app-polyfill": "^1.0.1",
-        "react-bootstrap": "^1.0.0-beta.9",
-        "react-dev-utils": "^9.0.1",
-        "react-dom": "^16.8.6",
-        "react-konva": "^16.8.7-3",
-        "react-native": "^0.60.5",
-        "react-redux": "^7.1.1",
-        "react-router-bootstrap": "^0.25.0",
-        "react-router-dom": "^5.0.1",
-        "request": "^2.88.0",
-        "request-promise": "^4.2.4",
-        "resolve": "1.10.0",
-        "sass-loader": "7.1.0",
-        "semver": "6.0.0",
-        "serverless-optimizer-plugin": "^2.5.1",
-        "serverless-plugin-existing-s3": "^2.4.0",
-        "set-value": "^3.0.1",
-        "style-loader": "0.23.1",
-        "terser-webpack-plugin": "1.2.3",
-        "ts-pnp": "1.1.2",
-        "url-loader": "1.1.2",
-        "uuid": "^3.3.2",
-        "uuid-base64": "^1.0.0",
-        "webpack": "4.29.6",
-        "webpack-dev-server": "3.2.1",
-        "webpack-manifest-plugin": "2.0.4",
-        "workbox-webpack-plugin": "4.2.0",
-        "zlib": "^1.0.5"
-    },
-    "scripts": {
-        "start": "node scripts/start.js",
-        "build": "node scripts/build.js",
-        "test": "cross-env TS_NODE_FILES=true mocha --require ts-node/register src/lambda/*.ts --timeout 10000",
-        "compile": "tsc src/lambda/*.ts",
-        "cov": "nyc --reporter=text npm run test",
-        "clear": "rm -rf src/lambda/*.js",
-        "doc": "typedoc --out docs src/lambda",
-        "predeploy": "npm run build",
-        "deploy": "gh-pages -d build",
-        "lint": "eslint src/lambda/*.ts"
-    },
-    "eslintConfig": {
-        "extends": "react-app"
-    },
-    "browserslist": {
-        "production": [
-            ">0.2%",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
-    },
-    "devDependencies": {
-        "@types/babel__core": "^7.1.3",
-        "@types/babel__template": "^7.0.2",
-        "@types/chai": "^4.2.1",
-        "@types/chai-as-promised": "^7.1.2",
-        "@types/file-type": "^10.9.1",
-        "@types/jest": "^25.1.4",
-        "@types/mocha": "^5.2.7",
-        "@types/node-gzip": "^1.1.0",
-        "@types/react": "^16.8.23",
-        "@types/react-dom": "^16.8.4",
-        "@types/react-router-bootstrap": "^0.24.5",
-        "@types/uuid": "^3.4.5",
-        "@typescript-eslint/eslint-plugin": "^2.1.0",
-        "@typescript-eslint/parser": "^2.1.0",
-        "aws-sdk-mock": "^4.5.0",
-        "babel-cli": "^6.26.0",
-        "babel-plugin-jsx-control-statements": "^4.0.0",
-        "cross-env": "^5.2.0",
-        "eslint": "^6.3.0",
-        "eslint-config-airbnb": "^18.0.1",
-        "eslint-plugin-import": "^2.18.2",
-        "eslint-plugin-jsx-a11y": "^6.2.3",
-        "eslint-plugin-react": "^7.14.3",
-        "eslint-plugin-react-hooks": "^1.7.0",
-        "mocha": "^6.2.0",
-        "serverless-offline": "^5.8.0",
-        "serverless-plugin-optimize": "^4.1.0-rc.1",
-        "ts-jest": "^25.2.1",
-        "ts-mocha": "^6.0.0",
-        "ts-node": "^8.3.0",
-        "typedoc": "^0.15.0"
-    },
-    "babel": {
-        "presets": [
-            "react-app"
-        ],
-        "plugins": [
-            "jsx-control-statements"
-        ]
-    }
+  "name": "arenatime-client",
+  "version": "0.1.0",
+  "homepage": "https://stdev17.github.io/arenatime",
+  "private": true,
+  "dependencies": {
+    "@babel/core": "^7.8.7",
+    "@svgr/webpack": "4.1.0",
+    "acorn": "^7.1.1",
+    "athena-client": "^2.5.1",
+    "athena-express": "^5.1.1",
+    "aws-sdk": "^2.635.0",
+    "axios": "^0.19.2",
+    "babel-eslint": "10.0.1",
+    "babel-jest": "^25.1.0",
+    "babel-loader": "8.0.5",
+    "babel-plugin-named-asset-import": "^0.3.6",
+    "babel-preset-react-app": "^9.1.1",
+    "bootstrap": "^4.4.1",
+    "camelcase": "^5.2.0",
+    "case-sensitive-paths-webpack-plugin": "2.2.0",
+    "chai": "^4.2.0",
+    "chai-as-promised": "^7.1.1",
+    "compression": "^1.7.4",
+    "css-loader": "2.1.1",
+    "dotenv": "6.2.0",
+    "dotenv-expand": "4.2.0",
+    "eslint-config-react-app": "^4.0.1",
+    "eslint-loader": "2.1.2",
+    "eslint-plugin-flowtype": "2.50.1",
+    "eslint-utils": "^1.4.3",
+    "fetch": "^1.1.0",
+    "file-loader": "3.0.1",
+    "file-type": "^12.4.2",
+    "form-data": "^2.5.1",
+    "fs": "0.0.1-security",
+    "fs-extra": "7.0.1",
+    "gh-pages": "^2.2.0",
+    "html-webpack-plugin": "4.0.0-beta.5",
+    "identity-obj-proxy": "3.0.0",
+    "is-mobile": "^2.2.1",
+    "is-wsl": "^1.1.0",
+    "istanbul-lib-instrument": "^3.3.0",
+    "jest": "^25.1.0",
+    "jest-environment-jsdom-fourteen": "0.1.0",
+    "jest-resolve": "24.7.1",
+    "jest-watch-typeahead": "0.3.0",
+    "js-combinatorics": "^0.5.5",
+    "konva": "^3.4.1",
+    "lodash": "^4.17.15",
+    "lodash.template": "^4.5.0",
+    "mini-css-extract-plugin": "0.5.0",
+    "mixin-deep": "^2.0.1",
+    "moment": "^2.24.0",
+    "multimatch": "^4.0.0",
+    "node-gzip": "^1.1.2",
+    "nth-prime": "^1.0.4",
+    "optimize-css-assets-webpack-plugin": "5.0.1",
+    "pnp-webpack-plugin": "1.2.1",
+    "postcss-flexbugs-fixes": "4.1.0",
+    "postcss-loader": "3.0.0",
+    "postcss-normalize": "7.0.1",
+    "postcss-preset-env": "6.6.0",
+    "postcss-safe-parser": "4.0.1",
+    "react": "^16.13.0",
+    "react-app-polyfill": "^1.0.6",
+    "react-bootstrap": "^1.0.0-beta.17",
+    "react-dev-utils": "^9.1.0",
+    "react-dom": "^16.13.0",
+    "react-konva": "^16.8.7-4",
+    "react-native": "^0.60.6",
+    "react-redux": "^7.2.0",
+    "react-router-bootstrap": "^0.25.0",
+    "react-router-dom": "^5.1.2",
+    "request": "^2.88.2",
+    "request-promise": "^4.2.5",
+    "resolve": "1.10.0",
+    "sass-loader": "7.1.0",
+    "semver": "6.0.0",
+    "serialize-javascript": "^3.0.0",
+    "serverless-optimizer-plugin": "^2.5.1",
+    "serverless-plugin-existing-s3": "^2.4.0",
+    "set-value": "^3.0.1",
+    "style-loader": "0.23.1",
+    "terser-webpack-plugin": "^2.3.5",
+    "ts-pnp": "1.1.2",
+    "url-loader": "1.1.2",
+    "uuid": "^3.4.0",
+    "uuid-base64": "^1.0.0",
+    "webpack": "4.29.6",
+    "webpack-dev-server": "3.2.1",
+    "webpack-manifest-plugin": "2.0.4",
+    "workbox-webpack-plugin": "4.2.0",
+    "zlib": "^1.0.5"
+  },
+  "scripts": {
+    "start": "node scripts/start.js",
+    "build": "node scripts/build.js",
+    "test": "jest",
+    "compile": "tsc src/lambda/*.ts",
+    "cov": "nyc --reporter=text npm run test",
+    "clear": "rm -rf src/lambda/*.js",
+    "doc": "typedoc --out docs src/lambda",
+    "predeploy": "npm run build",
+    "deploy": "gh-pages -d build",
+    "lint": "eslint src/lambda/*.ts"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "airbnb",
+      "prettier"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "devDependencies": {
+    "@types/babel__core": "^7.1.6",
+    "@types/babel__template": "^7.0.2",
+    "@types/chai": "^4.2.10",
+    "@types/chai-as-promised": "^7.1.2",
+    "@types/file-type": "^10.9.1",
+    "@types/jest": "^25.1.4",
+    "@types/node-gzip": "^1.1.0",
+    "@types/react": "^16.9.23",
+    "@types/react-dom": "^16.9.5",
+    "@types/react-router-bootstrap": "^0.24.5",
+    "@types/uuid": "^3.4.8",
+    "@typescript-eslint/eslint-plugin": "^2.22.0",
+    "@typescript-eslint/parser": "^2.22.0",
+    "aws-sdk-mock": "^4.5.0",
+    "babel-cli": "^6.26.0",
+    "babel-plugin-jsx-control-statements": "^4.0.0",
+    "cross-env": "^5.2.1",
+    "eslint": "^6.8.0",
+    "eslint-config-airbnb": "^18.0.1",
+    "eslint-plugin-import": "^2.20.1",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-react": "^7.19.0",
+    "eslint-plugin-react-hooks": "^1.7.0",
+    "serverless-offline": "^5.12.1",
+    "serverless-plugin-optimize": "^4.1.0-rc.1",
+    "ts-jest": "^25.2.1",
+    "ts-mocha": "^6.0.0",
+    "ts-node": "^8.6.2",
+    "typedoc": "^0.15.8"
+  },
+  "babel": {
+    "presets": [
+      "react-app"
+    ],
+    "plugins": [
+      "jsx-control-statements"
+    ]
+  }
 }
