@@ -1,7 +1,4 @@
-import aws = require('aws-sdk');
-aws.config.update({region: 'ap-northeast-2'});
-const dyn = new aws.DynamoDB();
-let index = 'matches';
+import { dynamoQuery } from '../util/dynamo';
 
 /** 람다 핸들러 함수
  * @param event http request에 인자를 담아주세요
@@ -12,27 +9,28 @@ export const handler = async (event: any, context: any): Promise<any> => {
 
   const params = {
     TableName: 'comment-table',
-    IndexName: index,
+    IndexName: 'matches',
     ExpressionAttributeValues: {
-      ':Id': {S: req}
+      ':Id': { S: req },
     },
     KeyConditionExpression: 'matchId = :Id',
     ProjectionExpression: 'commentId, charName, memo, uploadedDate',
-    ScanIndexForward: true
+    ScanIndexForward: true,
   };
 
-  const res: any = await dyn.query(params).promise()
-    .then(data => {
+  const res: any = await dynamoQuery(params)
+    .then((data: any) => {
       const response = {
         statusCode: 200,
         body: JSON.stringify({
           message: data,
-          runtime: context
+          runtime: context,
         }),
         headers: {
+          'Content-Type': 'application/json',
           'Access-Control-Allow-Origin': 'https://stdev17.github.io',
           'Access-Control-Allow-Credentials': true,
-        }
+        },
       };
       return response;
     })
@@ -41,24 +39,16 @@ export const handler = async (event: any, context: any): Promise<any> => {
         statusCode: 400,
         body: JSON.stringify({
           message: 'Failed Comment Read',
-          runtime: context
+          runtime: context,
         }),
         headers: {
+          'Content-Type': 'application/json',
           'Access-Control-Allow-Origin': 'https://stdev17.github.io',
           'Access-Control-Allow-Credentials': true,
-        }
+        },
       };
       return response;
     });
 
-    return res;
-
-}
-
-/** 유닛 테스트에 호출되는 함수 */
-async function test (event: any, context: any, _index: string): Promise<any> {
-  index = _index;
-  return await handler(event, context);
-}
-
-export default test;
+  return res;
+};
